+++ demo/lib/main.dart
@@ -54,6 +54,7 @@ import 'screen/feature/pages_list_screen.dart';
 import 'screen/feature/frozen_rows_screen.dart';
 import 'screen/feature/scrollbars.dart';
 import 'screen/feature/row_wrapper_screen.dart';
+import 'screen/feature/multiitems_delegate_demo_screen.dart';
 
 void main() {
   runApp(const MyApp());
@@ -147,6 +148,8 @@ class MyApp extends StatelessWidget {
         LoadingOptionsScreen.routeName: (context) =>
             const LoadingOptionsScreen(),
         RowWrapperScreen.routeName: (context) => const RowWrapperScreen(),
+        MultiItemsDelegateDemoScreen.routeName: (context) =>
+            const MultiItemsDelegateDemoScreen(),
       },
       theme: ThemeData(
         primaryColor: TrinaGridExampleColors.primaryColor,